backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
python-multipart
